requirements.txt
paho-mqtt
PyYAML
streamlit
pandas
plotly
influxdb-client
pytest
python-dotenv